refactor: make processHandler flatter
